Add one-off script to fix bug in February's survey
I was attempting to get stats from the previous survey and use those if the current
survey had 0-value stats, but this failed.  Ran this script to
interactively fix the 14 or so trainer submissions for February.
Two trainers had two February surveys, so for those I hand edited this
script to use values from the 3rd most recent instead of 2nd most recent
survey.